requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database and ORM
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Vector Database
qdrant-client==1.7.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# LLM and Embeddings
openai==1.3.7
anthropic==0.7.8
sentence-transformers==2.2.2
langchain==0.0.340
tiktoken==0.5.2

# Document Processing
pypdf2==3.0.1
python-docx==1.1.0
python-magic==0.4.27

# Configuration and Environment
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# HTTP and Async
httpx==0.25.2
aiofiles==23.2.1
redis==5.0.1

# Streamlit Frontend
streamlit==1.28.2
streamlit-authenticator==0.2.3

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Monitoring and Logging
structlog==23.2.0